Add files via upload
LivelinessBackend solution is a backend microservice.
In the context of software architecture, a microservice is a small, independent process that communicates with other microservices over a network to fulfill a larger system’s goals. Each microservice is self-contained and should implement a single business capability.

So, if LivelinessBackend is a microservice, it would be responsible for a specific function or feature in your larger application, and it would communicate with other services to contribute to the overall functionality of the system.